+++ include/blickfeld/scanner.h
@@ -244,19 +244,30 @@ public:
 	std::shared_ptr<scanner::point_cloud_stream<protocol::data::Frame> > get_point_cloud_stream(const protocol::config::ScanPattern_Filter filter, const protocol::data::Frame reference_frame);
 
 	/**
-	 * > Introduced in BSL v2.15 and firmware v1.9
+	 * > Introduced in BSL v2.17
 	 *
 	 * Fetches point cloud frames from the device.
 	 * This call already boots up the device.
 	 * It fails if the device is in an errored state.
 	 * Delete the shared reference to stop the stream.
 	 *
-	 * @param filter Filter points and returns by point attributes during the post-processing on the device.
-	 * @param reference_frame Frame representing the desired data. To request a field, set it to any value (also in submessages). For a repeated field, add at least one element.
-	 * @param reference_subscription Template subscription. Enabled extenstions from this parameter are used to open the stream.
+	 * @param algorithms List of algorithms to enable for this stream.
 	 * @return Shared pointer of stream instance. Use the recv_frame method to get Point Cloud Frames.
 	 */
-	std::shared_ptr<scanner::point_cloud_stream<protocol::data::Frame> > get_point_cloud_stream(const protocol::config::ScanPattern_Filter filter, const protocol::data::Frame reference_frame, const protocol::stream::Subscribe::PointCloud extend_subscription);
+	std::shared_ptr<scanner::point_cloud_stream<protocol::data::Frame> > get_point_cloud_stream(const std::vector<protocol::config::Algorithm> algorithms);
+
+	/**
+	 * > Introduced in BSL v2.17
+	 *
+	 * Fetches point cloud frames from the device.
+	 * This call already boots up the device.
+	 * It fails if the device is in an errored state.
+	 * Delete the shared reference to stop the stream.
+	 *
+	 * @param extend_subscription Reference subscription. These parameters can enable on-device algorithms, filter values and customize the data fields in the point cloud output.
+	 * @return Shared pointer of stream instance. Use the recv_frame method to get Point Cloud Frames.
+	 */
+	std::shared_ptr<scanner::point_cloud_stream<protocol::data::Frame> > get_point_cloud_stream(const protocol::stream::Subscribe::PointCloud extend_subscription);
 
 #ifdef BSL_RECORDING